+++ modules/core.lua
@@ -1,5 +1,5 @@
 EZOGroupFrames = EZOGroupFrames or {}
 
 EZOGroupFrames.ADDON_NAME = "EZOGroupFrames"
-EZOGroupFrames.ADDON_VERSION = "0.1.4"
+EZOGroupFrames.ADDON_VERSION = "0.1.5"
 EZOGroupFrames.AUTHOR = "@Zuriplayer"